// src/api/types.gen.ts
export type ClusterReference = {
  /**
   * Cluster Id
   *
   * The identifier of the cluster/canonical entity that is considered equivalent to the
   * subject entity mention that an `EntityMentionResolutionResponse` refers to.
   *
   */
  cluster_id: string
  /**
   * Confidence Score
   *
   * A 0-1 value of how confident the ERE is about the equivalence between the subject entity mention
   * and the target canonical entity.
   *
   */
  confidence_score: number
  /**
   * Similarity Score
   *
   * A 0-1 score representing the pairwise comparison between a mention and a cluster (likely
   * based on a representative representation).
   *
   */
  similarity_score: number
}
// ...
export type DecisionSummary = {
  /**
   * Id
   */
  id: string
  about_entity_mention: EntityMentionPreview
  current_placement: ClusterReference
  /**
   * Created At
   */
  created_at: string
  /**
   * Updated At
   */
  updated_at?: string | null
}
// ...
export type EntityMentionIdentifier = {
  /**
   * Source Id
   *
   * The ID or URI of the ERS client that originated the request. This identifies an application or a
   * person accessing the ERS system.
   *
   */
  source_id: string
  /**
   * Request Id
   *
   * A string representing the unique ID of the request made to the ERS system. In general, this is unique
   * only within the scope of the source and the entity type, ie, within `sourceId` and `entityType`.
   *
   * Moreover, this is **not** the same as `ereRequestId`, which instead, is internal to the ERE and is
   * used to match responses to requests.
   *
   */
  request_id: string
  /**
   * Entity Type
   *
   * A string representing the entity type (based on CET). This is typically a URI.
   *
   * Note that this is at this level, and not at `EntityMention`, since, as said above,
   * it's needed to identify the entity, even when its content is not present. For the same
   * reason, it's used both for `EREResolutionRequest` and `EREResolutionResponse` messages.,
   *
   */
  entity_type: string
}

/**
 * EntityMentionPreview
 *
 * Lightweight entity mention projection for display.
 */
export type EntityMentionPreview = {
  identified_by: EntityMentionIdentifier
  /**
   * Parsed Representation
   */
  parsed_representation?: {
    [key: string]: unknown
  } | null
}